+++ new/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: clacaill <clacaill@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 13:43:59 by clacaill          #+#    #+#             */
-/*   Updated: 2022/11/07 14:45:03 by clacaill         ###   ########.fr       */
+/*   Updated: 2022/11/07 18:02:48 by clacaill         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strnstr(const char *big, const char *little, size_t len)
 {
-	int	i;
+	size_t	i;
 	int	j;
 
 	i = 0;
